fix: fallback to populate menus on page load
Menu data is supposed to be retrieved from craft during site build and inserted into the static pages, which will initialize the vuex store. However, if nuxt doesn't find a page and build it into the static site, it can load as a SPA without populating the menus. This PR ensures that menus get populated in that case. (We also need to figure out why pages aren't getting populated).